refact: TestTags moved to feature_auth
