+++ app/src/main/java/com/imaan/store/feature_auth/presentation/utils/TestTags.kt
@@ -1,4 +1,4 @@
-package com.imaan.sharedtest.utils.test_tags.login
+package com.imaan.store.feature_auth.presentation.utils
 
 object TestTags {
     const val subtitle = "subtitle"
@@ -7,4 +7,6 @@ object TestTags {
     const val loginButton = "loginBtn"
     const val signUpText = "signUpBtn"
     const val dontHaveAccount = "dontHaveAccount"
+
+    const val signUpPage = "signUpPage"
 }
