--- sharedTest/src/main/java/com/imaan/sharedtest/utils/test_tags/signup/TestTags.kt
@@ -1,6 +1,0 @@
-package com.imaan.sharedtest.utils.test_tags.signup
-
-object TestTags {
-
-    const val signUpPage = "signUpPage"
-}
